chore: execute http request when key is not null

+++ app/src/main/java/com/demo/pagination/feature/NetworkPagingMediator.kt
@@ -74,13 +74,13 @@ class NetworkPagingMediator(
                                 highestIndexInfo.index - currentPage
                             }
                         }
-                    } ?: 1 // this should only happen on first load, anchor is only null at the first load
+                    }
                 }
                 // In this example, you never need to prepend, since REFRESH
                 // will always load the first page in the list. Immediately
                 // return, reporting end of pagination.
                 LoadType.PREPEND ->
-                    return MediatorResult.Success(endOfPaginationReached = true)
+                    null
                 LoadType.APPEND -> {
 
                     // The logic that determines what is the last item could be anything but in our
@@ -95,48 +95,52 @@ class NetworkPagingMediator(
                 }
             }
 
-            // Suspending network load via Retrofit. This doesn't need to be
-            // wrapped in a withContext(Dispatcher.IO) { ... } block since
-            // Retrofit's Coroutine CallAdapter dispatches on a worker
-            // thread.
-            val response = tmdbService.getPopularShows(page = index)
+            if (index != null) {
+                // Suspending network load via Retrofit. This doesn't need to be
+                // wrapped in a withContext(Dispatcher.IO) { ... } block since
+                // Retrofit's Coroutine CallAdapter dispatches on a worker
+                // thread.
+                val response = tmdbService.getPopularShows(page = index)
 
-            showsDb.withWriteTransaction<Unit> {
+                showsDb.withWriteTransaction<Unit> {
 
-                when (loadType) {
+                    when (loadType) {
 //                    userDao.deleteShowsWrapper()
-                    LoadType.PREPEND -> TODO()
-                    LoadType.APPEND, LoadType.REFRESH -> {
-                        if (
-                            loadType == LoadType.REFRESH &&
-                            highestIndexInfo != null &&
-                            lowestIndexInfo != null
-                        ) {
-                            // this means data already exists and we need to delete it
-                            tvShowDao.deleteAllTvShows()
-                            pageIndexesDao.deleteAll()
-                        }
+                        LoadType.PREPEND -> TODO()
+                        LoadType.APPEND, LoadType.REFRESH -> {
+                            if (
+                                loadType == LoadType.REFRESH &&
+                                highestIndexInfo != null &&
+                                lowestIndexInfo != null
+                            ) {
+                                // this means data already exists and we need to delete it
+                                tvShowDao.deleteAllTvShows()
+                                pageIndexesDao.deleteAll()
+                            }
 
-                        response.body()?.results?.also {
-                            tvShowDao.insertAll(
-                                Array(it.size) { index ->
-                                    it[index].toTvShowQuery()
-                                }
-                            )
-                        }
+                            response.body()?.results?.also {
+                                tvShowDao.insertAll(
+                                    Array(it.size) { index ->
+                                        it[index].toTvShowQuery()
+                                    }
+                                )
+                            }
 
-                        pageIndexesDao.upsert(
-                            PageIndexesEntity(
-                                key = HIGHEST_INDEXES_ID,
-                                index = index
+                            pageIndexesDao.upsert(
+                                PageIndexesEntity(
+                                    key = HIGHEST_INDEXES_ID,
+                                    index = index
+                                )
                             )
-                        )
 
-                        if (lowestIndexInfo == null) {
-                            PageIndexesEntity(
-                                key = LOWEST_INDEXES_ID,
-                                index = index
-                            )
+                            if (lowestIndexInfo == null) {
+                                pageIndexesDao.upsert(
+                                    PageIndexesEntity(
+                                        key = LOWEST_INDEXES_ID,
+                                        index = index
+                                    )
+                                )
+                            }
                         }
                     }
                 }
